+++ source/silver/src/silver/infrastructure/helpers/environment_variable_helper.py
